feat: add assert last event
