+++ Cargo.toml
@@ -1,11 +1,11 @@
 [workspace]
 resolver = "2"
 members = [
-  "crates/ink-sandbox",
-  "crates/drink/drink",
-  "crates/drink/drink-cli",
-  "crates/drink/drink/test-macro",
-  "crates/pop-drink",
+    "crates/ink-sandbox",
+    "crates/drink/drink",
+    "crates/drink/drink-cli",
+    "crates/drink/drink/test-macro",
+    "crates/pop-drink",
 ]
 exclude = ["crates/drink/examples"]
 
@@ -30,7 +30,7 @@ proc-macro2 = { version = "1" }
 quote = { version = "1" }
 ratatui = { version = "0.21.0" }
 scale = { package = "parity-scale-codec", version = "3.6.9", features = [
-  "derive",
+    "derive",
 ] }
 scale-info = { version = "2.10.0" }
 serde_json = { version = "1.0" }
@@ -54,6 +54,6 @@ sp-runtime-interface = { version = "28.0.0", features = ["std"] }
 # Local
 drink = { path = "crates/drink/drink" }
 ink_sandbox = { path = "crates/ink-sandbox" }
-pop-api = { git = "https://github.com/r0gue-io/pop-node.git", branch = "main" }
+pop-api = { git = "https://github.com/r0gue-io/pop-node.git", branch = "chungquantin/feat-assert_last_event" }
 pop-drink = { path = "crates/pop-drink" }
 pop-runtime-devnet = { git = "https://github.com/r0gue-io/pop-node.git", branch = "main" }
